# backend/openapi.yaml
openapi: 3.0.3
info:
  title: RavelGo API
  version: "0.1.0"
  description: >
    Unified REST API for the RavelGo rider, driver, and admin apps.
    Authenticated with Amazon Cognito access tokens; role checks are enforced
    via Cognito group membership (`Rider`, `Driver`, `Admin`) — see
    `src/middleware/auth.ts`. There is also a WebSocket endpoint
    (`wss://<host>/ws?token=<access token>`) for live trip-status and
    driver-location push, not representable in OpenAPI 3.0 — see
    `docs/realtime-architecture.md` for its wire protocol. Every REST
    endpoint below has an HTTP-polling equivalent, so a client never
    strictly needs the WebSocket connection.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /car-paddy:
    post:
      tags: [Car Paddy]
      summary: Submit a Car Paddy (vehicle license renewal) request (Driver)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plateNumber]
              properties:
                plateNumber: { type: string }
      responses:
        "201":
          description: Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CarPaddyRequest" }
    get:
      tags: [Car Paddy]
      summary: List all Car Paddy requests (Admin)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated requests
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCarPaddyRequest" }
  /car-paddy/{id}:
    patch:
      tags: [Car Paddy]
      summary: Approve/reject a Car Paddy request (Admin)
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [APPROVED, REJECTED] }
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CarPaddyRequest" }
...
  /car-paddy/{id}/document-url:
    get:
      tags: [Car Paddy]
      summary: Signed S3 URL for a Car Paddy request's supporting document (Admin)
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Short-lived signed URL
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/cash-reconciliation:
    get:
      tags: [Admin]
      summary: Cash owed by each driver against what has been remitted (Admin)
      responses:
        "200":
          description: Cash reconciliation rows
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/cash-remittances:
    post:
      tags: [Admin]
      summary: Record a cash remittance handed in by a driver (Admin, cash:write)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [driverId, amount]
              properties:
                driverId: { type: string }
                amount: { type: number }
                note: { type: string }
      responses:
        "201":
          description: Cash remittance
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Cognito access token
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Validation or bad request error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Caller lacks the required Cognito group
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    CarPaddyRequest:
      type: object
      properties:
        id: { type: string }
        driverId: { type: string }
        plateNumber: { type: string }
        status: { type: string, enum: [SUBMITTED, IN_REVIEW, APPROVED, REJECTED] }
...
    PaginatedCarPaddyRequest:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/CarPaddyRequest" } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_ERROR
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_SERVER_ERROR
                - RATE_LIMIT_EXCEEDED
                - BAD_REQUEST
                - PAYMENT_ERROR
                - DATABASE_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              nullable: true
              description: Field-level validation errors
              items:
                type: object
                properties:
                  field: { type: string, description: 'Field name or path (e.g., "email", "address.zipCode")' }
                  message: { type: string, description: Error message for this field }
                  code: { type: string, nullable: true, description: 'Error code (e.g., "too_small", "invalid_email")' }
            timestamp:
              type: string
              format: date-time
              description: When the error occurred (ISO 8601)